+++ src/components/ui/components/TitleBar.jsx
@@ -1,8 +1,8 @@
 import React from 'react';
-import { BiSolidLeaf } from 'react-icons/bi';
 import { IoClose } from 'react-icons/io5';
 import useTitleBar from '../hooks/useTitleBar';
 import ThemeSwitch from './theme/ThemeSwitch';
+import Image from 'next/image';
 
 export default function TitleBar() {
     const { windowClose } = useTitleBar();
@@ -12,7 +12,12 @@ export default function TitleBar() {
             <div className='absolute top-0 h-[62px] select-none z-50 w-full'>
                 <div className='flex justify-between items-center h-full text-titletext' data-tauri-drag-region>
                     <div className='flex justify-center items-center gap-1 px-2 h-full w-[62px]'>
-                        <BiSolidLeaf fontSize={40} />
+                        <Image
+                            src={'/logo.png'}
+                            width={40}
+                            height={40}
+                            alt='logo'
+                        />
                     </div>
 
                     <ThemeSwitch />