+++ changelog.md
@@ -1,4 +1,6 @@
 [![Downloads](https://img.shields.io/github/downloads/zevnda/steam-game-idler/1.6.1/total?style=for-the-badge&logo=github&color=137eb5)](https://github.com/zevnda/steam-game-idler/releases/download/1.6.1/Steam.Game.Idler_1.6.1_x64_en-US.msi)
 
 ## Changelog
-- Fixed an issue where the information div was hidden while running `card farming`
+- New logo/icon for both the SGI and SteamUtility windows
+  - If anyone is good at making custom logos/icons, let me know
+- Fixed the height of the draggable area of the `titlebar` when on the `settings` screen